Refactor out inner class for buffered enumeration
